LayoutTests/platform/wk2/http/tests/security/mixedContent/redirect-https-to-http-iframe-in-main-frame-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
CONSOLE MESSAGE: [blocked] The page at about:blank was not allowed to display insecure content from http://127.0.0.1:8080/security/mixedContent/resources/boring.html.

main frame - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
This test opens a window that loads an insecure iframe (via a tricky redirect). We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying content that can be controlled by an active network attacker.
